Update deploy and autodiscovery scripts
Autodiscovery now ensures that only Running Servers are added
to the list (these are the only ones returned, but this is used
as a showcase).

Deploy now pulls all branches from the remote, uses a port-dependent
PID file, and reloads the app instead of terminating it if it's already
running.